set up router - refactored app component into home component - created a /user URL - moved the add event button from NavBar to the EventList - made the logo in the NavBar a link to the “home page”
