+++ src/components/event_list.js
@@ -51,6 +51,12 @@ class EventList extends Component {
       if (this.eventItems.length > 0) {
         return (
           <div id="event-menu">
+            <div className="add-event-btn-container">
+              <button className="add-event-plus" type="button" onClick={this.props.toggleAddEvent}>
+                Add Event
+                <img id="plus" src="./../../icon_set_1/plus.png" role="presentation" />
+              </button>
+            </div>
             <input id="search-bar" type="text" value={this.state.searchString} onChange={this.handleChange} placeholder="Type here..." />
             {this.eventItems}
           </div>
@@ -60,6 +66,12 @@ class EventList extends Component {
     // Case of no matching events.
     return (
       <div id="event-none">
+        <div className="add-event-btn-container">
+          <button className="add-event-plus" type="button" onClick={this.props.toggleAddEvent}>
+            Add Event
+            <img id="plus" src="./../../icon_set_1/plus.png" role="presentation" />
+          </button>
+        </div>
         <input id="search-bar" type="text" value={this.state.searchString} onChange={this.handleChange} placeholder="Type here..." />
         <text className="warning-msg">
           No Matching Events. <br />
