+++ src/components/nav_bar.js
@@ -1,15 +1,15 @@
 // nav_bar.js
 import React from 'react';
+import { Link } from 'react-router';
 
 const NavBar = (props) => {
   return (
     <div id="nav-bar">
-      <img id="logo" src="./../../images/dartmap.png" role="presentation" />
-      <h1 className="app-name">dartmap</h1>
-      <button type="button" onClick={props.toggleAddEvent}>
-        Add Event
-        <img id="plus" src="./../../icon_set_1/plus.png" role="presentation" />
-      </button>
+      <Link to="/" className="logo-link nav-btn">
+        <img id="logo" src="./../../images/dartmap.png" role="presentation" />
+        <h1 className="app-name">dartmap</h1>
+      </Link>
+      <Link to="/user" className="user-link nav-btn">Profile</Link>
     </div>
   );
 };
