+++ src/components/user_page.js
@@ -1,0 +1,11 @@
+// user_page.js
+
+import React from 'react';
+
+const UserPage = (props) => {
+  return (
+    <div> I am the user profile page.</div>
+  );
+};
+
+export default UserPage;
